SERVER-14668 Fix compile break on 32-bit and Clang
Introduced by commit e671a2c31f6b555254e4ad8265f66578f8c805aa

+++ src/mongo/db/catalog/collection.cpp
@@ -441,7 +441,7 @@ namespace mongo {
         return _recordStore->dataSize( txn );
     }
 
-    uint64_t Collection::getIndexSize(OperationContext* opCtx,
+    size_t Collection::getIndexSize(OperationContext* opCtx,
                                       BSONObjBuilder* details,
                                       int scale) {
